Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/sandbox-in-meta-tag-ignored-expected.txt

CONSOLE MESSAGE: The Content Security Policy directive 'sandbox' is ignored when delivered via an HTML meta element.
ALERT: PASS executed script.
This tests that the Content Security Policy sandbox directive is ignored when contained in a policy defined via an HTML meta element.
